Use C99 conversion specifiers in the runtime.
C99 introduced additional conversion specifiers for use with the printf and
scanf family of functions. Specifically, the 'z' size specifier for size_t
values and the 't' size specifier for ptrdiff_t values. Historically, we have
not used them because the C99 support in a certain C compiler was lacking.
Since this is no longer the case, use them. This removes the need of a bunch
of casts, some of which were incorrect on 64-bit Windows.

runtime/mercury_accurate_gc.c:
runtime/mercury_memory_zones.c:
runtime/mercury_stacks.c:
runtime/mercury_wrapper.c:
    As above.

+++ runtime/mercury_stacks.c
@@ -1,7 +1,7 @@
 // vim: ts=4 sw=4 expandtab ft=c
 
 // Copyright (C) 1998-2001, 2003-2006, 2008, 2010-2011 The University of Melbourne.
-// Copyright (C) 2014, 2016, 2018 The Mercury team.
+// Copyright (C) 2014, 2016, 2018, 2026 The Mercury team.
 // This file is distributed under the terms specified in COPYING.LIB.
 
 // This file contains code for printing statistics about stack frame sizes,
@@ -91,8 +91,8 @@ MR_print_stack_frame_stats(void)
         det_frame_total_size);
     fprintf(fp, "average size of a det stack frame:   %.3f\n",
         det_frame_total_size / det_frame_count);
-    fprintf(fp, "max size of det stack:               %ld\n",
-        (long) (MR_det_frame_max - MR_CONTEXT(MR_ctxt_detstack_zone)->min));
+    fprintf(fp, "max size of det stack:               %td\n",
+        (MR_det_frame_max - MR_CONTEXT(MR_ctxt_detstack_zone)->min));
     fprintf(fp, "\n");
 
     fprintf(fp, "number of non stack frames created:  %.0f\n",
@@ -101,8 +101,8 @@ MR_print_stack_frame_stats(void)
         non_frame_total_size);
     fprintf(fp, "average size of a non stack frame:   %.3f\n",
         non_frame_total_size / non_frame_count);
-    fprintf(fp, "max size of non stack:               %ld\n",
-        (long) (MR_non_frame_max - MR_CONTEXT(MR_ctxt_nondetstack_zone)->min));
+    fprintf(fp, "max size of non stack:               %td\n",
+        (MR_non_frame_max - MR_CONTEXT(MR_ctxt_nondetstack_zone)->min));
     fprintf(fp, "-------------------------------------------\n");
 
     MR_checked_fclose(fp, MR_STACK_FRAME_STATS);
